+++ packages/apolloschurchapp/src/client/httpLink.js
@@ -5,6 +5,6 @@ import { APP_DATA_URL } from 'react-native-dotenv';
 let uri = APP_DATA_URL;
 console.log({ uri }); // huh?
 // Android's emulator requires localhost network traffic to go through 10.0.2.2
-if (Platform.OS === 'ios') uri = uri.replace('localhost', '192.168.200.17');
+if (Platform.OS === 'android') uri = uri.replace('localhost', '10.0.2.2');
 
 export default createHttpLink({ uri });
